requirements.txt
# Для Задачи 1
transformers>=4.20.0
torch>=1.9.0
sentence-transformers>=2.2.0
numpy>=1.21.0
requests>=2.25.0
scikit-learn>=1.1.0
feedparser>=6.0.0 # Для RSS
joblib>=1.2.0 # Для сохранения моделей

# Для Задачи 2
opencv-python>=4.5.0 # Для работы с видео и изображениями
torchvision>=0.10.0 # Для предобученных моделей (CNN, Vision Transformer)
decord>=0.6.0 # Альтернатива OpenCV для более быстрого чтения видео
transformers # Уже есть, но нужен для CLIP, BERT
pillow>=8.0.0 # Для работы с изображениями
pandas>=1.3.0 # Для удобства работы с метаданными